server/requirements.txt
Flask
Flask-Cors
tensorflow
numpy
Pillow
python-dotenv